user_bot/requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
aiosqlite~=0.21.0
aiogram~=3.19.0
APScheduler~=3.11.0
aiohttp~=3.11.14
yookassa~=3.5.0